requirements.txt
Django==4.2.7
cryptography==41.0.5
requests==2.31.0
djangorestframework==3.14.0
PyJWT==2.8.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
python-dotenv==1.0.0
django-cors-headers==4.3.0